Rename rule name to ts-imports
