+++ src/index.ts
@@ -1,10 +1,10 @@
 'use strict'
 
 import noRelativePathImports from './rules/no-relative-path-imports'
-import preferTsPathsImports from './rules/prefer-ts-paths-imports'
+import tsImports from './rules/ts-imports'
 
 // Import all rules in lib/rules
 module.exports.rules = {
-  'prefer-ts-paths-imports': preferTsPathsImports,
+  'ts-imports': tsImports,
   'no-relative-path-imports': noRelativePathImports,
 }
